refactor(integration): extract src/modules/integration module
- Crée src/modules/integration/{index,bus,events,auth}.ts
- Extrait les helpers auth (requireIntegrationAccess, isIntegrationMember,
  IntegrationScope) de src/lib/auth vers le module
- Extrait les builders d'email et runInactivityNotifications vers
  services/family-service.ts — supprime la duplication entre
  cron/route.ts et cron/integration-inactivity/route.ts
- Enregistre le module dans src/lib/registry.ts
- Ajoute la règle de frontière no-integration-imports-other-modules
  dans .dependency-cruiser.cjs
- Mise à jour de tous les imports dans src/app/ vers @/modules/integration
- Aucun changement de comportement (lint:boundaries + typecheck verts)

Refs #357
