+++ src/app/(auth)/integration/stats/page.tsx
@@ -1,4 +1,5 @@
-import { requireAuth, getCurrentChurchId, requireIntegrationAccess } from "@/lib/auth";
+import { requireAuth, getCurrentChurchId } from "@/lib/auth";
+import { requireIntegrationAccess } from "@/modules/integration";
 import { prisma } from "@/lib/prisma";
 import StatsView from "./StatsView";
