+++ src/app/api/integration/families/route.ts
@@ -1,5 +1,5 @@
 import { successResponse, errorResponse, ApiError } from "@/lib/api-utils";
-import { requireIntegrationAccess } from "@/lib/auth";
+import { requireIntegrationAccess } from "@/modules/integration";
 
 const FAMILIES_API_URL =
   process.env.FAMILIES_API_URL ?? "https://familles.iccrennes.fr";
