+++ src/app/api/integration/requests/[id]/route.ts
@@ -1,9 +1,8 @@
 import { prisma } from "@/lib/prisma";
 import { successResponse, errorResponse, ApiError } from "@/lib/api-utils";
-import { requireIntegrationAccess } from "@/lib/auth";
 import { logAudit } from "@/lib/audit";
-import { sendEmail } from "@/lib/email";
 import { z } from "zod";
+import { requireIntegrationAccess, notifyBergerAssigned } from "@/modules/integration";
 
 export async function GET(
   _request: Request,
@@ -195,78 +194,19 @@ export async function PATCH(
 
     // Notifier le berger à l'affectation
     if (notifyBergerId) {
-      const berger = await prisma.user.findUnique({
-        where: { id: notifyBergerId },
-        select: { id: true, name: true, email: true },
+      const appUrl = process.env.APP_URL ?? process.env.AUTH_URL ?? process.env.NEXTAUTH_URL ?? "http://localhost:3000";
+      await notifyBergerAssigned({
+        bergerId: notifyBergerId,
+        firstName: req.firstName,
+        lastName: req.lastName,
+        requestId: id,
+        familyName: updated.assignedFamilyName,
+        appUrl,
       });
-      if (berger) {
-        await prisma.notification.create({
-          data: {
-            userId: berger.id,
-            type: "INTEGRATION_ASSIGNED",
-            title: "Nouvelle demande d'intégration",
-            message: `${req.firstName} ${req.lastName} vous a été affecté${updated.assignedFamilyName ? ` (${updated.assignedFamilyName})` : ""}.`,
-            link: `/admin/integration/requests/${id}`,
-          },
-        }).catch(() => {});
-
-        if (berger.email) {
-          await sendEmail({
-            to: berger.email,
-            subject: "Nouvelle demande d'intégration vous a été affectée",
-            html: buildBergerNotifEmail({
-              bergerName: berger.name ?? berger.email,
-              firstName: req.firstName,
-              lastName: req.lastName,
-              familyName: updated.assignedFamilyName,
-              requestId: id,
-            }),
-          }).catch(() => {});
-        }
-      }
     }
 
     return successResponse(updated);
   } catch (error) {
     return errorResponse(error);
   }
-}
-
-function buildBergerNotifEmail(params: {
-  bergerName: string;
-  firstName: string;
-  lastName: string;
-  familyName: string | null;
-  requestId: string;
-}): string {
-  const { bergerName, firstName, lastName, familyName, requestId } = params;
-  const appUrl = process.env.APP_URL ?? process.env.AUTH_URL ?? process.env.NEXTAUTH_URL ?? "http://localhost:3000";
-  return `<!DOCTYPE html>
-<html lang="fr">
-<head><meta charset="UTF-8"><meta name="viewport" content="width=device-width,initial-scale=1"></head>
-<body style="margin:0;padding:0;background:#f9fafb;font-family:Arial,sans-serif;">
-  <div style="max-width:560px;margin:40px auto;background:#fff;border-radius:12px;overflow:hidden;box-shadow:0 1px 4px rgba(0,0,0,.08)">
-    <div style="background:#5E17EB;padding:28px 32px 20px">
-      <h1 style="margin:0;color:#fff;font-size:20px;font-weight:700">Nouvelle demande d'intégration</h1>
-    </div>
-    <div style="padding:28px 32px">
-      <p style="margin:0 0 14px;color:#111827;font-size:15px">Bonjour ${bergerName},</p>
-      <p style="margin:0 0 16px;color:#374151;font-size:14px;line-height:1.6">
-        Une demande d'intégration vient de vous être affectée :
-      </p>
-      <div style="background:#f5f3ff;border-left:4px solid #5E17EB;padding:12px 16px;border-radius:0 8px 8px 0;margin:0 0 20px">
-        <p style="margin:0;color:#111827;font-size:15px;font-weight:600">${firstName} ${lastName}</p>
-        ${familyName ? `<p style="margin:4px 0 0;color:#6b7280;font-size:13px">Famille : ${familyName}</p>` : ""}
-      </div>
-      <a href="${appUrl}/admin/integration/requests/${requestId}"
-         style="display:inline-block;background:#5E17EB;color:#fff;text-decoration:none;padding:10px 20px;border-radius:8px;font-size:14px;font-weight:600">
-        Voir la demande →
-      </a>
-    </div>
-    <div style="background:#f9fafb;padding:14px 32px;border-top:1px solid #e5e7eb">
-      <p style="margin:0;color:#9ca3af;font-size:11px">Notification automatique Koinonia.</p>
-    </div>
-  </div>
-</body>
-</html>`;
 }
